refactor: moved pydantic prompts to its own class
