+++ src/ragas/prompt/__init__.py
@@ -1,5 +1,6 @@
-from .base import BasePrompt, BoolIO, PydanticPrompt, StringIO, StringPrompt
+from .base import BasePrompt, BoolIO, StringIO, StringPrompt
 from .mixin import PromptMixin
+from .pydantic_prompt import PydanticPrompt
 
 __all__ = [
     "BasePrompt",
